flatten_mirror.py: Add option to 'push' vs copy when flattening
We want a simple way to create a subset of the repositories with only the
branches we care about.  Using the new 'PUSH_NOT_COPY' option, if the
repository is not bare, we will push to a new empty repository vs copying
the existing directory.

We also adjust the processing to capture the base_url from the default.xml
so that when we adjust the json files we don't break the vcs_url entries.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>