branch_mirror.py: Add a new command to branch the existing mirror
usage: branch_mirror.py <branch> [--force]

This will parse the existing default.xml and then branch all non-bare
repositories with the selected branch.  If that branch already exists, use
the --force option to force the branch to move to a different commit.

The tool will also write out a list of layers that were modified to
branched-layers.list.  The format of this file is 'repository branch'.

This code does not push or otherwise affect any upstream repositories, this
needs to be done as a separate option.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>