flatten_mirror.py: Rework, add mode to subset items while copying/flattening
Use a new argument '--subset-mirror' to point to a text file that includes
the layer name and the special 'folder' to copy it to when subsetting.

[SKIP] is a special key word that skips the item.

'#' comments and blank lines are allowed.

Much of the other code was reordered/refactored to add support for subsetting.

It's important that everything is processed (even if the copy/push is skipped)
so that we get proper errors generated for items that are not in the subset
file.  This will help prevent a subset and missing files.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>