settings.py: Make it clear that the bitbake path is based on oe-core path
Adjust setup.py to enforce that bitbake path is based on oe-core path.  This
change was made to address issues when a mirror is split and the bitbake
directory is moved into a folder.  The path to bitbake, in the right folder,
is only able to be learned by using openembedded-core's path.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>