utils_setup.py: Cleanup error handling in fetch_url
The DNS resolution error, -2, needs to be handled outside of this function.
Some users of this may accept a DNS failure, while in others it is a
critical failure.

Also be sure to expand on the reason for various errors, so any previously
unknown exception can be diagnosed more quickly.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>