flatten_mirror: Adjust oe-core grouping to also include wrlinux-x
Issue: LIN10-4070

To make things easier to manage we've now set the policy that:

wrlinux-x
git-repo
bitbake (and associated files)
openembedded-core (layer)

will always be managed as a unit, under the subset entry for
openembedded-core.  This follows the existing practices, but ensures that
no matter how we break things up, they will be consistent in the future.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>