bin/layer_index.py: Add support to filter the objects by matching yp_compatible_version
Issue: US117235

In the current implementation, all the objects(distro, machine and
wrtemplate) would be listed by default when using the arguments
--list-{distros,machines}. But some of them may be
introduced by other vendor's layer and are not supported by Wind River.
So the results of these list arguments seem pretty confusion.
This patch tries to tackle this issue by leveraging the
yp_compatible_version tag. The yp_compatible_version in the layerindex
database is got from the value of LAYERSERIES_COMPAT setted in the
layer.conf. We can add "wrl" to the LAYERSERIES_COMPAT and make the
--list-xxx arguments only list the corresponding object in these layers.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>