data/samples/local.conf.sample: Update systemd/sysvinit
It is unfortunately not possible to setup the systemd/sysvinit as a template
because of the way the template system works and when the 'BACKFILL' pieces
are processed.

We could adjust the base system to allow it, but then we run the signficant
risk of breaking other non-wrlinux distributions.  So it's better to simply
expand the local.conf documentation with the full options.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>