bin/layer_index.py: Prevent loading mirror files not ending in .json
Issue: LIN1019-51

With a new README file, we can't just arbitrarily load all files, we need
to filter out files without .json extensions.

This follows the behavior in the existing path loading function, as well
as the functions that actually write the index files.  (.json is always
the added file extension.)

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>