setup: Support checking out by tag instead of branch
Issue: LINUXEXEC-1839

Some operations need to be split into a fetch and checkout in order to work
with tags instead of branches.

Define a rule to determine how to convert a branch into a tag name.  Since
the branch of a index will have the branch and not tag name.  The rules are:

A tag must start with 'vWRLINUX_'.  If it contains _RCPL or _UPDATE it is a
product tag.  Otherwise it's considered to be a CI tag.  And then must end in
a version, starting with '10.'.

For instance product tag, vWRLINUX_10_18_UPDATE0007, refers to product branch
WRLINUX_10_18_UPDATE0007.

While a CI tag, vWRLINUX_CI_10.19.35.0, refers to a CI branch
WRLINUX_CI.

Once the conversion is known, we adjust the layerindex branch information to
refer to the tag instead of branch.  (Tags always are prefixed by
'refs/tags/vWRLINUX' in order to be identified.)

Additionally the setup.sh will not only detect the tag, but will also detect
a checkout that is related to tag, but not actually on the tag and force the
user to select a specific branch or tag.  (This should only ever happen on
internal development.)  If this is detected, the user will get a message:

Checkout may be on a detached HEAD, this HEAD does not appear to
correspond to a specific, tag.  (It appears you may be working with
tag <tag>.  If this is correct, use
--base-branch=<tag> in the arguments to setup.sh.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>