setup.py: add ability to check required host tools
Issue: LINUXEXEC-2667

Check necessary host tools in setup process. The list is
from HOSTTOOLS setting in our project.

The host_requirements.py holds this list in 'hosttools' variable,
and this value is expected to be updated if autorebase detects
some change.

We also fall back to a fixed list and give user a warning if
something goes wrong.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>