setup.sh: Adjust slightly, remove call to setup_help.py
Issue: LINUXEXEC-2547

To avoid any dependencies in setup.sh on 'python', when the --help is used
we want to just call setup.py now.  However, if the user doesn't have python3
we want to warn them and avoid crashing.  This was accomplished by moving the
python checks out of the 'if help' block.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>